requirements.txt
flask 
pymongo 
flask-marshmallow 
marshmallow
python-dotenv
flask-jwt-extended